requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.1.0
Django==5.2.10
django-allauth==65.13.1
django-bootstrap-icons==0.9.0
django-bootstrap5==26.1
django-crispy-forms==2.5
django-environ==0.12.0
django-filter==25.2
django-htmx==1.27.0
django-tables2==2.8.0
idna==3.11
psycopg==3.3.2
psycopg-binary==3.3.2
pyaml==25.7.0
pycparser==2.23
PyJWT==2.10.1
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
uWSGI==2.0.31
